requirements.txt
python-dotenv==0.10.2
requests==2.21.0